#9 - Refactored Favourites screen.
